Add product for all collections
