+++ dask_expr/tests/test_collection.py
@@ -126,11 +126,6 @@ def test_setitem(pdf, df):
     assert_eq(df, pdf)
 
 
-def test_series_product(pdf, df):
-    assert_eq(pdf.x.product(), df.x.product())
-    assert_eq(pdf.y.product(), df.y.product())
-
-
 @xfail_gpu("https://github.com/rapidsai/cudf/issues/10271")
 def test_explode():
     pdf = lib.DataFrame({"a": [[1, 2], [3, 4]]})
@@ -178,109 +173,6 @@ def test_dask(pdf, df):
     z = (df.x + df.y).sum()
 
     assert assert_eq(z, (pdf.x + pdf.y).sum())
-
-
-@pytest.mark.parametrize(
-    "func",
-    [
-        M.max,
-        M.min,
-        M.any,
-        M.all,
-        M.sum,
-        M.prod,
-        M.count,
-        M.mean,
-        M.std,
-        M.var,
-        pytest.param(
-            M.idxmin, marks=xfail_gpu("https://github.com/rapidsai/cudf/issues/9602")
-        ),
-        pytest.param(
-            M.idxmax, marks=xfail_gpu("https://github.com/rapidsai/cudf/issues/9602")
-        ),
-        pytest.param(
-            lambda df: df.size,
-            marks=pytest.mark.xfail(reason="scalars don't work yet"),
-        ),
-    ],
-)
-def test_reductions(func, pdf, df):
-    result = func(df)
-    assert result.known_divisions
-    assert_eq(result, func(pdf))
-    result = func(df.x)
-    assert not result.known_divisions
-    assert_eq(result, func(pdf.x))
-    # check_dtype False because sub-selection of columns that is pushed through
-    # is not reflected in the meta calculation
-    assert_eq(func(df)["x"], func(pdf)["x"], check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "func",
-    [
-        M.max,
-        M.min,
-        M.any,
-        M.all,
-        lambda idx: idx.size,
-    ],
-)
-def test_index_reductions(func, pdf, df):
-    result = func(df.index)
-    assert not result.known_divisions
-    assert_eq(result, func(pdf.index))
-
-
-@pytest.mark.parametrize(
-    "func",
-    [
-        lambda idx: idx.index,
-        M.sum,
-        M.prod,
-        M.count,
-        M.mean,
-        M.std,
-        M.var,
-        M.idxmin,
-        M.idxmax,
-    ],
-)
-def test_unimplemented_on_index(func, pdf, df):
-    # Methods/properties of Series that don't exist on Index
-    with pytest.raises(AttributeError):
-        func(pdf.index)
-    with pytest.raises(AttributeError, match="'Index' object has no attribute '"):
-        func(df.index)
-
-
-def test_reduction_on_empty_df():
-    pdf = lib.DataFrame()
-    df = from_pandas(pdf)
-    assert_eq(df.sum(), pdf.sum())
-
-
-@pytest.mark.parametrize("axis", [0, 1])
-@pytest.mark.parametrize(
-    "skipna",
-    [
-        True,
-        pytest.param(
-            False, marks=xfail_gpu("cudf requires skipna=True when nulls are present.")
-        ),
-    ],
-)
-@pytest.mark.parametrize("ddof", [1, 2])
-def test_std_kwargs(axis, skipna, ddof):
-    pdf = lib.DataFrame(
-        {"x": range(30), "y": [1, 2, None] * 10, "z": ["dog", "cat"] * 15}
-    )
-    df = from_pandas(pdf, npartitions=3)
-    assert_eq(
-        pdf.std(axis=axis, skipna=skipna, ddof=ddof, numeric_only=True),
-        df.std(axis=axis, skipna=skipna, ddof=ddof, numeric_only=True),
-    )
 
 
 @pytest.mark.parametrize("func", ["cumsum", "cumprod", "cummin", "cummax"])
