+++ dask_expr/tests/test_reductions.py
@@ -2,10 +2,11 @@ from __future__ import annotations
 
 import numpy as np
 import pytest
+from dask.utils import M
 
 from dask_expr import from_pandas
 from dask_expr._util import DASK_GT_20231201
-from dask_expr.tests._util import _backend_library, assert_eq
+from dask_expr.tests._util import _backend_library, assert_eq, xfail_gpu
 
 # Set DataFrame backend for this module
 lib = _backend_library()
@@ -200,3 +201,107 @@ def test_value_counts_split_out_normalize(df, pdf):
     result = df.x.value_counts(split_out=2, normalize=True)
     expected = pdf.x.value_counts(normalize=True)
     assert_eq(result, expected)
+
+
+@pytest.mark.parametrize(
+    "func",
+    [
+        M.max,
+        M.min,
+        M.any,
+        M.all,
+        M.sum,
+        M.prod,
+        M.product,
+        M.count,
+        M.mean,
+        M.std,
+        M.var,
+        pytest.param(
+            M.idxmin, marks=xfail_gpu("https://github.com/rapidsai/cudf/issues/9602")
+        ),
+        pytest.param(
+            M.idxmax, marks=xfail_gpu("https://github.com/rapidsai/cudf/issues/9602")
+        ),
+        pytest.param(
+            lambda df: df.size,
+            marks=pytest.mark.xfail(reason="scalars don't work yet"),
+        ),
+    ],
+)
+def test_reductions(func, pdf, df):
+    result = func(df)
+    assert result.known_divisions
+    assert_eq(result, func(pdf))
+    result = func(df.x)
+    assert not result.known_divisions
+    assert_eq(result, func(pdf.x))
+    # check_dtype False because sub-selection of columns that is pushed through
+    # is not reflected in the meta calculation
+    assert_eq(func(df)["x"], func(pdf)["x"], check_dtype=False)
+
+
+@pytest.mark.parametrize(
+    "func",
+    [
+        M.max,
+        M.min,
+        M.any,
+        M.all,
+        lambda idx: idx.size,
+    ],
+)
+def test_index_reductions(func, pdf, df):
+    result = func(df.index)
+    assert not result.known_divisions
+    assert_eq(result, func(pdf.index))
+
+
+@pytest.mark.parametrize(
+    "func",
+    [
+        lambda idx: idx.index,
+        M.sum,
+        M.prod,
+        M.count,
+        M.mean,
+        M.std,
+        M.var,
+        M.idxmin,
+        M.idxmax,
+    ],
+)
+def test_unimplemented_on_index(func, pdf, df):
+    # Methods/properties of Series that don't exist on Index
+    with pytest.raises(AttributeError):
+        func(pdf.index)
+    with pytest.raises(AttributeError, match="'Index' object has no attribute '"):
+        func(df.index)
+
+
+def test_reduction_on_empty_df():
+    pdf = lib.DataFrame()
+    df = from_pandas(pdf)
+    assert_eq(df.sum(), pdf.sum())
+
+
+@pytest.mark.parametrize("axis", [0, 1])
+@pytest.mark.parametrize(
+    "skipna",
+    [
+        True,
+        pytest.param(
+            False, marks=xfail_gpu("cudf requires skipna=True when nulls are present.")
+        ),
+    ],
+)
+@pytest.mark.parametrize("ddof", [1, 2])
+def test_std_kwargs(axis, skipna, ddof):
+    pdf = lib.DataFrame(
+        {"x": range(30), "y": [1, 2, None] * 10, "z": ["dog", "cat"] * 15}
+    )
+    df = from_pandas(pdf, npartitions=3)
+    assert_eq(
+        pdf.std(axis=axis, skipna=skipna, ddof=ddof, numeric_only=True),
+        df.std(axis=axis, skipna=skipna, ddof=ddof, numeric_only=True),
+    )
